# docs/swagger.yaml
definitions:
  model.ExpireOption:
    enum:
    - 1
    - 2
    - 3
    - 4
    type: integer
    x-enum-varnames:
    - Expire6Hours
    - Expire12Hours
    - Expire1Day
    - Expire7Days
  model.ExpiredTypeResponse:
    properties:
      name:
        type: string
      type:
        type: integer
    type: object
  model.PasteRequest:
    properties:
      content:
        type: string
      expire:
        allOf:
        - $ref: '#/definitions/model.ExpireOption'
        enum:
        - 1
        - 2
        - 3
        - 4
      language:
        type: string
      title:
        type: string
    required:
    - content
    - expire
    - language
    - title
    type: object
  model.PasteResponse:
    properties:
      content:
        type: string
      created_at:
        type: string
      expires_at:
        type: string
      id:
        example: abc123
        type: string
      language:
        type: string
      title:
        type: string
    required:
    - language
    - title
    type: object
  model.PasteUpdateRequest:
    properties:
      content:
        type: string
      expire:
        $ref: '#/definitions/model.ExpireOption'
      language:
        type: string
      title:
        type: string
    type: object
  model.ResponseFormat:
    properties:
      data: {}
      error_message:
        type: string
      message:
        type: string
    type: object
info:
  contact: {}
  description: JsonJunk
  title: JsonJunk API
  version: "1.0"
paths:
  /api/v1/paste:
    post:
      consumes:
      - application/json
      description: 새로운 Paste 텍스트를 생성하고 저장합니다.
      parameters:
      - description: Paste Content
        in: body
        name: paste
        required: true
        schema:
          $ref: '#/definitions/model.PasteRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.ResponseFormat'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ResponseFormat'
      summary: Paste 생성
      tags:
      - pastes
  /api/v1/paste/{id}:
    delete:
      consumes:
      - application/json
      description: Paste 텍스트를 삭제합니다.
      parameters:
      - description: Paste ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.ResponseFormat'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ResponseFormat'
      summary: Paste 삭제
      tags:
      - pastes:test
    get:
      consumes:
      - application/json
      description: ID를 통해 저장된 Paste 내용을 조회합니다.
      parameters:
      - description: Paste ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.ResponseFormat'
            - properties:
                data:
                  $ref: '#/definitions/model.PasteResponse'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.ResponseFormat'
      summary: Paste 조회
      tags:
      - pastes
    put:
      consumes:
      - application/json
      description: Paste 텍스트를 업데이트 및 저장합니다.
      parameters:
      - description: Paste ID
        in: path
        name: id
        required: true
        type: string
      - description: Paste update Content
        in: body
        name: paste
        required: true
        schema:
          $ref: '#/definitions/model.PasteUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.ResponseFormat'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ResponseFormat'
      summary: Paste 업데이트
      tags:
      - pastes:test
  /api/v1/paste/list:
    get:
      consumes:
      - application/json
      description: Paste 목록 조회 요청
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.ResponseFormat'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.PasteResponse'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ResponseFormat'
      summary: Paste 목록 조회
      tags:
      - pastes
  /api/v1/paste/type:
    get:
      consumes:
      - application/json
      description: Paste expire type 조회합니다.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.ResponseFormat'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.ExpiredTypeResponse'
                  type: array
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.ResponseFormat'
      summary: Paste expire type 조회
      tags:
      - pastes:type
  /raw/{id}:
    get:
      consumes:
      - application/json
      description: 데이터를 조회한다.
      parameters:
      - description: Paste ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      summary: 데이터를 조회한다.
      tags:
      - api
swagger: "2.0"